use actual type for txIndex and clauseIndex in tracer.Context

+++ tracers/tracers.go
@@ -32,9 +32,9 @@ import (
 type Context struct {
 	BlockID     thor.Bytes32 // Hash of the block the tx is contained within (zero if dangling tx or call)
 	BlockTime   uint64       // Timestamp of the block the tx is contained within
-	TxIndex     int          // Index of the transaction within a block (zero if dangling tx or call)
+	TxIndex     uint64       // Index of the transaction within a block (zero if dangling tx or call)
 	TxID        thor.Bytes32 // ID of the transaction being traced (zero if dangling call)
-	ClauseIndex int          // Index of the clause within a transaction (zero if dangling call)
+	ClauseIndex uint32       // Index of the clause within a transaction (zero if dangling call)
 	State       *state.State
 }
